Added the ability to define the keys to use in the store. Added an action that can be called to set the new keys to be used.
Added the keylisteners in the instructions and stimuli component.

Added a callback via props to the components so that they can change the sate of the parent.

Added conditional rendering for the experiment index based on the components state.

Made the blackscreen callback to transition into the next state after 1 second has passed.